remaining scientific code moved to blocks and library folders
