+++ obi_one/scientific/blocks/extracellular_locations.py
@@ -1,11 +1,11 @@
 from obi_one.core.block import Block
 
 
-class ExtracellularLocationSet(Block):
+class ExtracellularLocations(Block):
     """Base class of extracellular locations."""
 
 
-class XYZExtracellularLocationSet(ExtracellularLocationSet):
+class XYZExtracellularLocations(ExtracellularLocations):
     xyz_locations: (
         tuple[tuple[float, float, float], ...] | list[tuple[tuple[float, float, float], ...]]
     ) = ((0.0, 0.0, 0.0),)
