+++ obi_one/scientific/blocks/stimulus.py
@@ -10,7 +10,7 @@ from pydantic import Field, NonNegativeFloat, PrivateAttr
 from obi_one.core.block import Block
 from obi_one.core.constants import _MIN_NON_NEGATIVE_FLOAT_VALUE, _MIN_TIME_STEP_MILLISECONDS
 from obi_one.core.exception import OBIONEError
-from obi_one.scientific.circuit.circuit import Circuit
+from obi_one.scientific.library.circuit import Circuit
 from obi_one.scientific.unions.unions_neuron_sets import NeuronSetReference
 from obi_one.scientific.unions.unions_timestamps import TimestampsReference
