--- obi_one/scientific/circuit/intracellular_location_sets.py
@@ -1,11 +1,0 @@
-from obi_one.core.block import Block
-
-
-class IntracellularLocationSet(Block):
-    """Base class of intracellular locations."""
-
-    neuron_ids: tuple[int, ...] | list[tuple[int, ...]]
-
-
-class SectionIntracellularLocationSet(IntracellularLocationSet):
-    section: str | list[str]
