--- obi_one/scientific/circuit_metrics/__init__.py
@@ -1,1 +1,0 @@
-from .circuit_metrics import get_circuit_metrics as get_circuit_metrics
