+++ obi_one/scientific/library/specified_morphology_locations.py
