+++ obi_one/scientific/tasks/connectivity_matrix_extraction.py
@@ -9,7 +9,7 @@ from obi_one.core.block import Block
 from obi_one.core.scan_config import ScanConfig
 from obi_one.core.single_config_mixin import SingleConfigMixin
 from obi_one.core.task import Task
-from obi_one.scientific.circuit.circuit import Circuit
+from obi_one.scientific.library.circuit import Circuit
 
 L = logging.getLogger(__name__)
