+++ obi_one/scientific/tasks/morphology_locations.py
@@ -17,7 +17,7 @@ from obi_one.core.scan_config import ScanConfig
 from obi_one.core.single_config_mixin import SingleConfigMixin
 from obi_one.core.task import Task
 from obi_one.database.reconstruction_morphology_from_id import ReconstructionMorphologyFromID
-from obi_one.scientific.morphology_locations.specified_morphology_locations import (
+from obi_one.library.specified_morphology_locations import (
     _PRE_IDX,
     _SEC_ID,
     _SEG_ID,
