+++ obi_one/scientific/unions/unions_neuron_sets.py
@@ -3,7 +3,7 @@ from typing import Annotated, Any, ClassVar
 from pydantic import Discriminator
 
 from obi_one.core.block_reference import BlockReference
-from obi_one.scientific.circuit.neuron_sets import (
+from obi_one.scientific.blocks.neuron_sets import (
     AllNeurons,
     CombinedNeuronSet,
     ExcitatoryNeurons,
